+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## New
 * Render notes as Markdown
 * Require login by default
+* Add stats for dropped purchases
 
 ## Improved
 * mark refunded purchases red on game overview